+++ pages/1 measurement_one.py
@@ -1,0 +1,5 @@
+import streamlit as st
+from classes import c
+
+st.write("The following kinematics files are valid:")
+st.write(c.kinematics)
